feat: add empty div
he reason for this change is that the movies array will initially be empty, and it will remain empty while the fetching process takes place in the background. Until the data is finally fetched, you don't want to display the message “The list is empty” since you don’t yet know whether the fetched array will be empty or not. A more accurate message to display would be “Loading…” But a simpler solution is to not display anything if the array is empty (hence why, here, an empty div has been rendered using <div className="main-view" />)