re-factor: pass onLoggenIn method
A method, onLoggedIn, will be passed as a prop with the same name to LoginView (note that this just happens to be the same name—it’s not a constraint). This method will update the user state of the MainView component and will be called when the user has successfully logged in.