re-factor: add propTypes
Here, you set the static propTypes property on MovieCard to an object that contains special values provided as utilities by prop-types. These values help specify how MovieCard's props should look.

The props object must include a movie object (shape({...}) means that it’s an object).
The movie prop (object) may contain a Title key; if it does, then it must be of type string.
The props object must contain onMovieClick and it must be a function.
Based on this, if t